move map and grid logic to PermitGrid
